backend/requirements.txt
Flask==2.2.2
boto3==1.26.90
python-dotenv==0.21.0
werkzeug==2.2.2
flask-cors==3.0.10